[WIP] Changed to NvAPIWrapper codebase
Found some core problems with the NVAPI code that I wrote that meant that the NVAPI functions weren't working correctly and were crashing. Swapped the codebase to a modified version of NvAPIWrapper.net code because Soroush Falahati is a friggin legend. He used classes rather than structs which turned out to be a much more sensible idea that my struct-based attempt.

Also changed the AMDLibrary and NVIDIALibrary display profile structures to explicitly state whether the NVIDIA or AMD libraries are in use. This will allow us to avoid situations where we accidentally try to apply an AMD config that doesn't exist just because we now detect an AMD video card in the PC. WIth the new code this just wouldn't happen.