[WIP] Tweaks to improve DPI data storage
I had to tweak the display format to allow us to store only the data we need from the DPI data we get from windows. This should allow us to set the data easier.

Also added a check to make sure that the DPI structure size hasn't changed, just in case windows decides to modify how this works. This is not a public API, so it may change at any time and we need to be alerted to that fact!