First working SetDisplayConfig
This git committ gets NV_DISP_SetDisplayConfig working. The problem was that SetDisplayConfig does not like any sourcemodeinfo supplied when setting the display config, and it complains with an NVAPI_INVALID_ARGUMENT error. Removing everything except the TargetsInfo and TargetInfoCount worked!

The downside is that SetDisplayConfig only sets the NVIDIA video card displays part ofthe layout. If there is an AMD video card, then it doesn't do anything else with it. It also doesn't apply the source mode settings, nor does it setup the target setting in the targetinfo->details part of teh settings we stored. So it's kinda useless!

It's actually better to just let WinLibrary do it's SetDisplayConfig as it covers all displays that windows knows about, including NVIDIA connected ones, but I am going to keep this in there to make sure that surround settings are pulled down correctly.