Fixed bug with mosaic removal
There was an error in the disassembling of the mosaic screens that was introduced in this cutover of the NVIDIALibrary from my struct-based NVImport back to code from Soroush Falahati which is class-based. My code was incorrectly trying to use two screens in a 1x1 grid, which can only support 1 screen. This is now fixed, and each screen gets it's own 1x1 grid.