First version that avoids NVIDIA dGPU power saving
This version has been tested in a laptop with NVIDIA Optimus where the NVIDIA GPU was unable to respond, but the new NVIDIAExportsDll.dll seems to prevent teh laptop dGPU from going to sleep. I've doine the same thing for AMD (AMDExportsDll.dll) but I do not have the hardware to test it!

This has been tested swapping three screens between surround, non-surround, on, off, etc and the code has been able to successfully handle all the display layouts thrown at it. It seems like it's fairly robust, but only getting my testing team to use it will I know for sure.