Attempting to refresh taskbars after a successful layout change
We've tried get windows to redraw the taskbars after each layout change. This sometimes works, but doesn't seem to work at all when changing to or from a NVIDIA Surround profile. Unfiortunately this is a real pain, as the only thing that works is to restart windows explorer :(

There doesn't appear to be anything I can do to affect this or make it better :(.