Disabled applying the older GDI when applying a new config
Removed the GDI code from the WinLibrary.SetActiveConfig. THis was done to remove any chance ttha the GDI code is impairing the ability of the CCD codebase to set up the display properly. We are still recording the GDI settings in all profiles so that I can turn it on in the future.

Also disable the Taskbar settings logic, as that just slows down the video libary and we're REALLY unlikely to make use oif the information. It's better just not to collect it TBH.