Added code to specifically try to restore windows taskbars
The code uses 3 different ways of trying to get taskbars back, and it checks to see if the taskbars are back before actually trying the next thing. The code works in increasing difficulty through the following levels:

1. The first level sends a SHCNE_ASSOCCHANGED windows message to windows to notify it that the screen has changed. This can sometimes notify windows explorer that it needs to redraw a taskbar.

2. The second level grabs the current windows display config using the CCD API, and then reapplys the current display layout.

3. The third level uses windows restart manager to restart windows explorer itself whilst still keeping any explorere settings or windows explorer windows running. This is way better than using taskkill to kill the windows explorer process and thenstarting again.

Hopefully this will all be hidden from the users, and they won't even notice that we're doing this extra work to keep the taskbards showing.