Move the video card exported variable logic to video library loading code
Both the AMD and NVIDIA video cards will go to sleep if they have not screens attached to them. This breaks the video library processing. I have developed C++ dlls that will inform the AMD or NVIDIA driver that it needs to stay awake and not goto to sleep while this application is running.

This means that any laptop using this could use more power as the video card won't go to sleep, but it also means that this application will continue to work as tthe user wants it to. This is a fair trade off in my opinion.