Rewrote the Windows SetDisplayConfig logic to try multiple times
Hopefully this will make things more reliable. We validate the display config first, then we attempt to apply the settings. If that fails we try a second time after a short delay. If that fails we try again just using the topology (it should get us 90% there) after a short delay.

Hopefully that will make things more reliable.