Updated WinLibrary to avoid power off of monitors
Also adjusted it so that there are 3 different ways that we try to wake the displays, and also added 1 sec delays between them to make sure the the monitors have finished the previous responses before they go on to the next step.

Also added way more logging to see more detail about what the display is reporting.