[WIP] Attempt to fix broken ADL2_Display_DisplayMapConfig_Get
ADL2_Display_DisplayMapConfig_Get seems to be broken in the way that it returns displayTargets array. It currently returns a filtered list of displays but the returned ADL_DISPLAY_TARGET objects all have the wrong DISPLAY_ID values - specifically the DisplayLogicalAdapterIndex and the DisplayPhysicalAdapterIndex. This is breaking the equality matching.