[WIP] First Working ADLX library loading and function list
YAY! Finally got the ADLXCSharpBin.dll working for all current functions of ADLX(v1.4 at the time of writing). Not all of them have been tested of course, but there are currently something like 1900 functions or so in there.

There is still a lot of time needed to work out the functions I have to call to emulate the same features as there is now, but this should hopefully avoid the issues that have started to creep in with ADL2.