Working AMD save functionality, and added empty struct constructors
Added the ability to save the AMD desktops and displays, but cannot get the eyefinity desktop information from the desktopservices. Need to try and figure out the way to do this.

Also went through the NVIDIALibrary , AMDLibray and WinLibrary and added the new c# feature of being able to use empty struct constructors. I've been wanting to use that for years!