Working Eyefinity Display capture
I ended up actually leaving the code in there commented out, as I did finally devise a way to get the Eyefinity Display object from the Disktop object, but it took a bit of work to determine. Then in the end I didn't actually need the infrmation about the display grid of the Eyefinity display, so I parked the code for later.