Temporarily disable IntelLibrary
I don't have any Intel hardware that is new enough to support IGCL. This means there is no way to debug things. I will therefore have to disable it for now, but keep it in the build so that I can get it to work in the future.

For now though, I need to disable it so that I can test the new ADLX Library with other users, and make sure that works. Then I can do a DisplayMagician release, and then work on the IGCL support at a later date.